cloud_gcp/cloud_function_sources/requirements.txt
# Function dependencies, for example:
# package>=version

firebase-admin>=4.3.0
google-cloud-pubsub>=1.5.0
google-cloud-iot>=1.0.0